refactor: split fsblockstore to separate file
